refactor: types, utilities, errors...
Refactored the entire Sendblue API client to Typescript. This meant defining the raw response types, and writing utilities to transform the types & values from "snake_case" to "camelCase" for the best DX. Some, but not all are implemented, and I will add more response shapes as I find use for them.

The types are briefly documented, providing a quick summary of the value without crowding the Intellisense with param and return tags that will likely change in the future.

Implemented a custom `Sendblue` error class for throwing context-rich errors, and changed how errors are handled within the `Sendblue.api` helper.

Added a ".gitignore" and "tsconfig.json", and a "dist" output directory for transpiled JS files and types. Added Prettier and a config.

Updated the "README.md" and the "package.json" to accomodate for the changes, correct the entry point for the package, add the types pointer, a build cmd for `tsc`, and version the package.